[Refactor] 외래키(JoinColumn) 이름 변경, File pakage 변경
