+++ src/main/java/travelplanrepo/domain/board/controller/BoardController.java
@@ -11,8 +11,8 @@ import travelplanrepo.domain.board.service.BoardService;
 import travelplanrepo.domain.itinerary.dto.PostItineraryDto;
 import travelplanrepo.domain.itinerary.entity.Itinerary;
 import travelplanrepo.global.security.argumentresolver.LoginAccountId;
-import travelplanrepo.domain.File.File;
-import travelplanrepo.domain.File.FileProcessor;
+import travelplanrepo.domain.File.domain.File;
+import travelplanrepo.domain.File.service.FileService;
 
 import java.io.IOException;
 import java.util.List;
@@ -28,19 +28,19 @@ public class BoardController {
     private String scheduleImgPath;
 
     private final BoardService boardService;
-    private final FileProcessor fileProcessor;
+    private final FileService fileService;
 
     @PostMapping("/board")
     public String postBoard(@LoginAccountId Long accountId, @ModelAttribute PostBoardDto postBoardDto) throws IOException {
         Board board = postBoardDto.toBoard();
-        File thumbnail = fileProcessor.storeFile(postBoardDto.getThumbnail(), boardImgPath);
+        File thumbnail = fileService.storeFile(postBoardDto.getThumbnail(), boardImgPath);
         board.setThumbnail(thumbnail);
 
         List<PostItineraryDto> itineraryList = postBoardDto.getItineraryList();
         List<Itinerary> boardItineraryList = board.getItineraryList();
         for (int i = 0; i < itineraryList.size(); i++) {
             PostItineraryDto postItineraryDto = itineraryList.get(i);
-            File img = fileProcessor.storeFile(postItineraryDto.getImg(), scheduleImgPath);
+            File img = fileService.storeFile(postItineraryDto.getImg(), scheduleImgPath);
 
             Itinerary itinerary = boardItineraryList.get(i);
             itinerary.setImg(img);
